requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core app dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-multipart
python-dotenv
openai
chromadb

# ML / NLP
torch==2.2.0
sentence-transformers
transformers

# Parsing PDFs
pdfminer.six
pdfplumber
pypdfium2

# Utils
scikit-learn
scipy
numpy
requests
tqdm

# Optional: testing, rich logging
pytest
rich
